The signals blueprint shares a double instead of a decimal
The workflow aggregate of the loan approval held the offered rate as a
BigDecimal. Camunda 7 converted such a top-level value to a double on its
way into the engine; that conversion is being removed, so the value would
become an object variable and, without a serialization format, a base64
blob in Cockpit.

No BPMN expression reads the attribute, and no blueprint configures a
serialization format. Rather than adding one for a single value, the
aggregate now shares a Double: an engine has a variable type for it, so
the rate stays a number an operator can read and an expression can
compare. The published rate stays a BigDecimal where it belongs, in the
application's own InterestRate table, and Service#applyInterestRate
converts at the line where application data becomes something the
workflow shares.

+++ loan-approval/src/main/java/blueprint/workflowmodule/loanapproval/Service.java
@@ -157,6 +157,12 @@ public class Service {
    * was broadcast after one had been stored.
    * </p>
    *
+   * <p>
+   * The published rate is a decimal and the aggregate holds a double, so the conversion
+   * happens here, where application data becomes something the workflow shares with the
+   * BPMS. {@code Aggregate#interestRate} says why.
+   * </p>
+   *
    * @param loanApproval The workflow's aggregate.
    */
   public void applyInterestRate(
@@ -169,7 +175,7 @@ public class Service {
                 + loanApproval.getLoanRequestId()
                 + "'"));
 
-    loanApproval.setInterestRate(publishedRate.getPercentage());
+    loanApproval.setInterestRate(publishedRate.getPercentage().doubleValue());
 
     log.info(
         "Loan approval '{}' is offered at {}%",
